+++ pkg/chatbot/playback.go
@@ -59,21 +59,21 @@ func (a *App) timewarpCmd(ctx context.Context, user *users.User, _ []string) {
 
 	// exit early if we're on OS X
 	if helpers.RunningOnDarwin() {
-		a.IRC.Say("Sorry, timewarp isn't available right now")
+		a.Chat.Say("Sorry, timewarp isn't available right now")
 		return
 	}
 
 	// rate-limit the number of times this can run
 	if !c.UserIsAdmin(user.Username) {
 		if time.Now().Sub(lastTimewarpTime) < 20*time.Second {
-			a.IRC.Say("Not yet; enjoy the moment!")
+			a.Chat.Say("Not yet; enjoy the moment!")
 			return
 		}
 	}
 
 	// only say this if the caller is not me
 	if !c.UserIsAdmin(user.Username) {
-		a.IRC.Say("Here we go...!")
+		a.Chat.Say("Here we go...!")
 	}
 
 	// do the timewarp
@@ -86,21 +86,21 @@ func (a *App) jumpCmd(ctx context.Context, user *users.User, params []string) {
 
 	// exit early if we're on OS X
 	if helpers.RunningOnDarwin() {
-		a.IRC.Say("Sorry, jump isn't available right now")
+		a.Chat.Say("Sorry, jump isn't available right now")
 		return
 	}
 
 	// rate-limit the number of times this can run
 	if !c.UserIsAdmin(user.Username) {
 		if time.Now().Sub(lastTimewarpTime) < 20*time.Second {
-			a.IRC.Say("Not yet; enjoy the moment!")
+			a.Chat.Say("Not yet; enjoy the moment!")
 			return
 		}
 	}
 
 	// exit if the user gave no args or too many
 	if len(params) == 0 || len(params) > 2 {
-		a.IRC.Say("Usage: !jump [state]")
+		a.Chat.Say("Usage: !jump [state]")
 		return
 	}
 
@@ -113,23 +113,23 @@ func (a *App) jumpCmd(ctx context.Context, user *users.User, params []string) {
 	// check to see if we even have footage for this state
 	if _, ok := err.(*terrors.NoFootageForStateError); ok {
 		msg := fmt.Sprintf("No footage for %s... yet! ;)", titlecaseState)
-		a.IRC.Say(msg)
+		a.Chat.Say(msg)
 		return
 	}
 	// check to see if there was an error finding a candidate video
 	if err != nil {
 		slog.ErrorContext(ctx, "error from finding random video for state", "err", err)
-		a.IRC.Say("Usage: !jump [state]")
+		a.Chat.Say("Usage: !jump [state]")
 		return
 	}
 	// tell VLC to play it
 	err = a.VLC.PlayFileInPlaylist(ctx, randomVid.File())
 	if err != nil {
 		slog.ErrorContext(ctx, "error from VLC client", "err", err)
-		a.IRC.Say("Usage: !jump [state]")
+		a.Chat.Say("Usage: !jump [state]")
 		return
 	}
-	a.IRC.Say(fmt.Sprintf("Jumping to %s...!", titlecaseState))
+	a.Chat.Say(fmt.Sprintf("Jumping to %s...!", titlecaseState))
 	// update the currently-playing video
 	a.Video.GetCurrentlyPlaying(ctx)
 	// show the flag for the state
@@ -145,14 +145,14 @@ func (a *App) skipCmd(ctx context.Context, user *users.User, params []string) {
 
 	// exit early if we're on OS X
 	if helpers.RunningOnDarwin() {
-		a.IRC.Say("Sorry, skip isn't available right now")
+		a.Chat.Say("Sorry, skip isn't available right now")
 		return
 	}
 
 	// rate-limit the number of times this can run
 	if !c.UserIsAdmin(user.Username) {
 		if time.Now().Sub(lastTimewarpTime) < 20*time.Second {
-			a.IRC.Say("Not yet; enjoy the moment!")
+			a.Chat.Say("Not yet; enjoy the moment!")
 			return
 		}
 	}
@@ -167,7 +167,7 @@ func (a *App) skipCmd(ctx context.Context, user *users.User, params []string) {
 		n, err = strconv.Atoi(params[0])
 		// if conversion fails or they give too many args
 		if err != nil || len(params) > 1 {
-			a.IRC.Say("Usage: !skip [num]")
+			a.Chat.Say("Usage: !skip [num]")
 			return
 		}
 	}
@@ -190,14 +190,14 @@ func (a *App) backCmd(ctx context.Context, user *users.User, params []string) {
 
 	// exit early if we're on OS X
 	if helpers.RunningOnDarwin() {
-		a.IRC.Say("Sorry, back isn't available right now")
+		a.Chat.Say("Sorry, back isn't available right now")
 		return
 	}
 
 	// rate-limit the number of times this can run
 	if !c.UserIsAdmin(user.Username) {
 		if time.Now().Sub(lastTimewarpTime) < 20*time.Second {
-			a.IRC.Say("Not yet; enjoy the moment!")
+			a.Chat.Say("Not yet; enjoy the moment!")
 			return
 		}
 	}
@@ -212,7 +212,7 @@ func (a *App) backCmd(ctx context.Context, user *users.User, params []string) {
 		n, err = strconv.Atoi(params[0])
 		// if conversion fails or they give too many args
 		if err != nil || len(params) > 1 {
-			a.IRC.Say("Usage: !back [num]")
+			a.Chat.Say("Usage: !back [num]")
 			return
 		}
 	}